refactor(markdown): split the module along its parse/flatten passes
2293 lines in one file, where the largest unit is now 519. The split
follows the pipeline rather than the vocabulary: source is parsed into a
width-independent intermediate form, then flattened against a concrete
width.

- `item` — the intermediate form both passes speak
- `parse` — pass one: pulldown-cmark events to `MdItem`s
- `flatten` — pass two: `MdItem`s to width-fitted lines
- `table` — table layout, a third of pass two on its own
- `stream` — `MarkdownState` and the settled-prefix cache
- `image` — the `ImageResolver` seam and cell sizing
- `view` — the `Markdown` view

That boundary is the reason both caches work, so it is now written down
in the markdown spec: the settled-prefix cache can hold parsed blocks
across frames only because they carry no width, and a resize re-flattens
without re-tokenizing.

No public API change — the submodules are private and
`components::markdown` exports exactly what it did before, verified
item-by-item against the previous file. Tests move with the code they
cover, with the shared render helpers in a `testutil` module.

Instruction counts *improved*: 1.4-2.1% fewer on every markdown path
(`reflow.mixed` past the -2% tolerance), because smaller modules
re-partition codegen units and change what gets inlined — the same effect
that cost 3.7-5.5% when the crate grew in an earlier change, running the
other way. Baseline re-blessed so the gate protects the new numbers.
