+++ crates/tuika-codeformatters/benches/iai-baseline.json
@@ -2,7 +2,7 @@
   "_comment": "iai-callgrind instruction-count baseline. Deterministic and machine-independent for a fixed toolchain + libc. Regenerate with `python3 benches/check_iai.py --update` and commit alongside the code change that shifted the counts.",
   "tolerance_pct": 2.0,
   "benchmarks": {
-    "highlight.python": 94920407,
-    "highlight.rust": 291151508
+    "highlight.python": 94920813,
+    "highlight.rust": 291157409
   }
 }
